Refactor laynet

+++ utils/io.py
@@ -11,6 +11,6 @@ def save_nii(V, path, fstr=None):
             fstr += '.nii.gz'
         path = os.path.join(path, fstr)
     
-    V = V.astype(np.uint8)
+    # V = V.astype(np.uint8)
     img = nib.Nifti1Image(V, np.eye(4))
     nib.save(img, path)
